Cleaned up types
So we are not doing `.to_s` and `.to_i` after each other. no need anymore.

+++ src/components/domains/difference_component.cr
@@ -1,6 +1,6 @@
 class DifferenceComponent < BaseComponent
-  needs now : Int32
-  needs before : Int32
+  needs now : Int64
+  needs before : Int64
   needs reverse : Bool = false
 
   def render
